Separate and Refactor Transaction List Header - Refactor Transaction List header connector to just give primitive types to the component - Remove unused and props and styles in Transaction List Header - Create new utils functions getDenomination and getDefaultDenomination - Make WalletList also use getDenomination util function - Refactor and clean Transaction List Scene of unused props styles and components
